Overwrite single method from motmetrics and otherwise use pip package
