--- python-sdk/nuscenes/eval/tracking/motmetrics/__init__.py
@@ -1,6 +1,0 @@
-from .mot import MOTAccumulator
-import nuscenes.eval.tracking.motmetrics.lap
-import nuscenes.eval.tracking.motmetrics.metrics
-
-# Needs to be last line
-__version__ = '1.1.3'
